moving project_info.py to front side
