+++ gpt_automation/project_info.py
